RJSF refactor FormApp to separate SiP (#6880)
